requirements.txt
web3==5.29.1
eth-utils==1.10.0
eth-rlp==0.2.1
py-solc-x==1.1.1
Django==4.0.5
Pillow==9.1.1
python-dotenv==0.20.0